+++ src/main/java/com/lucendar/strm/common/strm/StrmCtrlReq.java
@@ -1,0 +1,8 @@
+package com.lucendar.strm.common.strm;
+
+public interface StrmCtrlReq {
+
+    boolean isLive();
+    String getReqId();
+    byte getCtrl();
+}